+++ Sources/IOSSupport/Common/GlobalAction.swift
@@ -22,8 +22,9 @@ public final class GlobalAction {
     public var didSetTargetLanguage: PublishRelay<TranslationLanguage> = .init()
 
     public var didEditWord: PublishRelay<UUID> = .init()
-    public var didDeleteWord: PublishRelay<Word> = .init()
+    public var didDeleteWords: PublishRelay<[Word]> = .init()
     public var didAddWord: PublishRelay<Void> = .init()
+    public var didMarkWordsAsMemorized = PublishRelay<Void>()
 
     public var didResetWordList: PublishRelay<Void> = .init()